requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0

# CLIP
open-clip-torch>=2.24.0

# API and backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.31.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
